Modularizando y cambiando de clase algunas funciones.
